Updated... this is more a brainstorm/braindump
There's a lot here right now that needs to be condensed into a set of commands, documentation for their use, a specification for programming (binary protocol), and implementation of a queue mechanism without storing anything on disk (staying in RAM)... unless there's encrypted storage?? Try to not overcomplicate it, dummy. Baby steps without making bad design choices that later get refactored in a great way.